Links to places people should report issues
Want to encourage everyone to submit corrections and ask questions on whatever channel(s) they feel most comfortable with. Part of the spirit of the challenge is feedback, mentoring, and encouragement. Obviously we hope there aren't glaring, massive problems with the API as speced in the test folder, but we'd rather accept feedback than run the challenge from a proverbial ivory tower.